requirements.txt
Django
gunicorn
django-heroku
django-allauth
requests
zoomus
django-tempus-dominus